Added some proper interaction and proper rotation
After looking up how openGL does it, it appears I was missing one or two matrices. In openGL it is proj,view,model. I had a broken proj and molded my view with the model. After putting the right numbers in the right matrices it worked somewhat. 

My math is still somewhat wrong though because I add rotations like this: A = AR, but instead it should be A = RA, but doing so messes up the offset gained by the 4th dimensional part of the matrices. So I am fully aware that something is wrong, not sure what specifically. But hey, it works.